feat(core): split ZK logic into dedicated verifiers module and update tests
