+++ gateway-contract/contracts/core_contract/src/lib.rs
@@ -2,20 +2,14 @@
 
 pub mod events;
 pub mod types;
+pub mod zk_verifiers;
 
 use soroban_sdk::{
     contract, contracterror, contractimpl, contracttype, panic_with_error, vec, Address, BytesN,
-    Env, Vec,
+    Env,
 };
 use types::ResolveData;
-
-#[contracttype]
-#[derive(Clone, Debug)]
-pub struct Groth16Proof {
-    pub pi_a: Vec<BytesN<32>>,
-    pub pi_b: Vec<Vec<BytesN<32>>>,
-    pub pi_c: Vec<BytesN<32>>,
-}
+use zk_verifiers::{verify, Groth16Proof};
 
 #[contracterror]
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
@@ -30,7 +24,6 @@ pub struct Contract;
 #[derive(Clone)]
 pub enum DataKey {
     Resolver(BytesN<32>),
-    VerifyingKey,
 }
 
 #[contracterror]
@@ -41,19 +34,6 @@ pub enum ResolverError {
 
 #[contractimpl]
 impl Contract {
-    // Acceptance Criteria: verify_proof implemented
-    pub fn verify_proof(_env: Env, _proof: Groth16Proof, public_inputs: Vec<BytesN<32>>) -> bool {
-        // Logic check: Ensure we have exactly 1 public input (the commitment)
-        if public_inputs.len() != 1 {
-            return false;
-        }
-
-        // Instruction Budget Optimization:
-        // In a real scenario, this is where the BN254 pairing check happens.
-        // For this implementation, we validate the proof structure exists.
-        _proof.pi_a.len() == 1 && _proof.pi_b.len() == 1 && _proof.pi_c.len() == 1
-    }
-
     pub fn register_resolver(
         env: Env,
         commitment: BytesN<32>,
@@ -63,8 +43,7 @@ impl Contract {
     ) {
         let public_inputs = vec![&env, commitment.clone()];
 
-        // Acceptance Criteria: Invalid proofs rejected with clear error
-        if !Self::verify_proof(env.clone(), proof, public_inputs) {
+        if !verify(env.clone(), proof, public_inputs) {
             panic_with_error!(&env, ZkError::InvalidProof);
         }
 
@@ -74,8 +53,21 @@ impl Contract {
         env.storage().persistent().set(&key, &data);
     }
 
+    pub fn set_memo(env: Env, commitment: BytesN<32>, memo_id: u64) {
+        let key = DataKey::Resolver(commitment);
+        let mut data = env
+            .storage()
+            .persistent()
+            .get::<DataKey, ResolveData>(&key)
+            .unwrap_or_else(|| panic_with_error!(&env, ResolverError::NotFound));
+
+        data.memo = Some(memo_id);
+        env.storage().persistent().set(&key, &data);
+    }
+
     pub fn resolve(env: Env, commitment: BytesN<32>) -> (Address, Option<u64>) {
         let key = DataKey::Resolver(commitment);
+
         match env.storage().persistent().get::<DataKey, ResolveData>(&key) {
             Some(data) => (data.wallet, data.memo),
             None => panic_with_error!(&env, ResolverError::NotFound),
